+++ frontend/src/components/topbanner/index.js
@@ -4,6 +4,8 @@ import { jsx, css } from '@emotion/core'
 import { Query } from 'react-apollo'
 import { Container } from '../container'
 import { LanguageSwitcher } from '../../LanguageSwitcher'
+import { Trans } from '@lingui/macro'
+import { PhaseBanner } from '../phase-banner'
 import { GET_LANGUAGE_QUERY } from '../../utils/queriesAndMutations'
 import rcmpbrandingeng from '../../images/rcmpbrandingeng.svg'
 import rcmpbrandingfre from '../../images/rcmpbrandingfre.svg'
